backend/requirements.txt
Django>=4.2,<5.0
django-tenants>=3.5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
channels>=4.0.0
channels-redis>=4.1.0
django-cors-headers>=4.0.0
gunicorn>=21.2.0
uvicorn>=0.23.0
qrcode>=7.4.0
Pillow>=10.0.0

requests
